week8/requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
transformers==4.30.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
streamlit==1.25.0
torch==2.0.1
langchain==0.0.235
openai==0.27.8
python-dotenv==1.0.0
plotly==5.15.0
seaborn==0.12.2
matplotlib==3.7.1
